requirements.txt
# API
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# AI / LLM
openai>=1.40.0

# Data
duckdb==1.0.0
pandas==2.2.2
openpyxl==3.1.5
pyarrow==17.0.0

# Visualización (Dashboard Agent)
plotly==5.22.0
matplotlib==3.9.0
kaleido==0.2.1

# Auth / Azure
msal==1.30.0
azure-identity==1.17.1
azure-monitor-ingestion==1.0.0

# Utils
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
httpx==0.27.0

# Testing
pytest==8.3.2
pytest-asyncio==0.23.8
